Add regression test for heuristic memory policy
Regression test covering the problem reported in #221.

The regression was that the caller gave up ownership also in the case of
`const ptr`, which is not the original behavior of cppyy, i.e. the
behavior that ROOT came to expect.

The fix that makes this test pass is provided in this PR:
https://github.com/wlav/CPyCppyy/pull/22